refactor(money): BigDecimal on Order and OrderItem
Slice 2 of the money migration. The pricing pipeline has computed in Money
(exact BigDecimal, scale 2) since PR #128, but the result was widened back to
double on the way into the order — so the exactness ended one step short of the
database, and orders.total_amount stayed a DOUBLE PRECISION column holding a
binary float where a count of cents belonged.

- Order.totalAmount / discountAmount / shippingCost and
  OrderItem.orderedProductPrice / discount are BigDecimal, mapped
  @Column(precision = 12, scale = 2). OrderDTO and OrderItemDTO follow, so the
  exact figure reaches the API instead of being rounded off at the boundary.
- V24 alters the five columns to NUMERIC(12,2). The ROUND in the USING clause is
  a no-op for every value the application wrote — those were already rounded to
  the cent — and a safety net for any legacy row carrying float noise.
- PriceBreakdown's accessors return Money rather than double. Callers that still
  need a double now say so with toDouble(), and each of those calls marks a
  boundary the next slice removes: the cart summary, the shipping calculator's
  thresholds, ReturnRequest.refundAmount. currentTotal() is gone — runningTotal()
  is the same thing once it returns Money.
- PaymentAttempt.expectedTotal is a Money, so StripePaymentGateway calls
  Money.toCents() on the same value the order carries rather than on a double
  recovered from it. That comparison is what decides whether a payment is
  accepted for an order.
- OrderRepository.getTotalRevenue() returns BigDecimal, because a SUM over a
  NUMERIC column comes back as one. ReturnServiceImpl was casting that result
  straight to (Double) — a ClassCastException waiting for the first return
  request listed after this change.

One visible behaviour change: the activity log now reads "placed for $99.90"
rather than "$99.9". BigDecimal keeps the scale it was given; double printed the
trailing zero away.

The outbox is unaffected in flight. Its payload is an OrderDTO serialised as
JSON, and a stored "totalAmount": 25.0 parses back into a BigDecimal, so events
written before this change still deliver.

Test (Testcontainers): the five columns are numeric and not double precision; an
order for 3 x 84.99 round-trips as exactly 254.97, equals the sum of its own
lines, and yields the 25497 cents the gateway is held to; and a hundred rows of
0.07 sum to exactly 7.00 — that sum is what the revenue reports are built from,
and it is the promise a float column could not make.

Still double: Product, Cart, and the entities behind them. Slice 3 next.

+++ ecom-backend/src/main/java/com/ecommerce/project/service/NotificationService.java
@@ -2,11 +2,12 @@ package com.ecommerce.project.service;
 
 import com.ecommerce.project.payload.NotificationDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface NotificationService {
 
-    void notifyAdminNewOrder(Long orderId, String customerEmail, Double totalAmount);
+    void notifyAdminNewOrder(Long orderId, String customerEmail, BigDecimal totalAmount);
 
     void notifyUserOrderStatusChanged(Long orderId, String email, String newStatus);
